Feature /scrum 84 reservation jpa  (#46)
* WIP entity

* WIP update type

* WIP jpa repository

* WIP update type

* WIP cancel, pickup

* WIP dirty checking

* Feature querydsl

* update type(Long) and annotation

* reservation jpa test

* reservation jpa test

* Delete src/test/java/org/nmfw/foodietree/domain/store/service/StoreApprovalServiceTest.java

* update type

* fix entity

+++ src/main/java/org/nmfw/foodietree/domain/customer/dto/resp/MyPageReservationDetailDto.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @Slf4j
 @Builder
 public class MyPageReservationDetailDto {
-    private int reservationId;
+    private Long reservationId;
     private String customerId;
     private String nickname;
     private LocalDateTime reservationTime; // 고객이 예약한 시간
